move DetasilViewOverlay to it's own folder
